refactor(api): deprecate route POST /schooling-registration-dependent-users/password-update
in favor of POST /sco-organization-learners/password-update
