+++ api/tests/acceptance/application/organization-learner-dependent-user-controller_test.js
@@ -116,96 +116,6 @@ describe('Acceptance | Route | organization-learner-dependent-user', function ()
     });
   });
 
-  describe('POST /api/schooling-registration-dependent-users/password-update', function () {
-    let organizationId;
-    let options;
-
-    beforeEach(async function () {
-      organizationId = databaseBuilder.factory.buildOrganization({ type: 'SCO', isManagingStudents: true }).id;
-      const userId = databaseBuilder.factory.buildUser().id;
-      databaseBuilder.factory.buildMembership({ organizationId, userId });
-
-      await databaseBuilder.commit();
-
-      options = {
-        method: 'POST',
-        url: '/api/schooling-registration-dependent-users/password-update',
-        headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
-        payload: {
-          data: {
-            attributes: {
-              'organization-id': organizationId,
-              'schooling-registration-id': null,
-            },
-          },
-        },
-      };
-    });
-
-    it('should return a 200 status after having successfully updated the password', async function () {
-      // given
-      const userId = databaseBuilder.factory.buildUser.withRawPassword().id;
-      const organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({
-        organizationId,
-        userId,
-      }).id;
-      options.payload.data.attributes['schooling-registration-id'] = organizationLearnerId;
-
-      await databaseBuilder.commit();
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return a 404 status when organizationLearner does not exist', async function () {
-      // given
-      options.payload.data.attributes['schooling-registration-id'] = 1;
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(404);
-    });
-
-    it("should return a 404 status when organizationLearner's userId does not exist", async function () {
-      // given
-      const organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({
-        organizationId,
-        userId: null,
-      }).id;
-      options.payload.data.attributes['schooling-registration-id'] = organizationLearnerId;
-
-      await databaseBuilder.commit();
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(404);
-    });
-
-    it('should return a 403 status when user does not belong to the same organization as organizationLearner', async function () {
-      // given
-      const wrongOrganization = databaseBuilder.factory.buildOrganization();
-      const organizationLearnerWithWrongOrganization = databaseBuilder.factory.buildOrganizationLearner({
-        organizationId: wrongOrganization.id,
-      });
-      await databaseBuilder.commit();
-
-      options.payload.data.attributes['schooling-registration-id'] = organizationLearnerWithWrongOrganization.id;
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(403);
-    });
-  });
-
   describe('POST /api/schooling-registration-dependent-users/recover-account', function () {
     it('should return a 200 status and student information for account recovery', async function () {
       // given
