+++ api/tests/acceptance/application/sco-organization-learner-controller_test.js
@@ -2680,4 +2680,184 @@ describe('Acceptance | Controller | sco-organization-learners', function () {
       });
     });
   });
+
+  describe('POST /api/schooling-registration-dependent-users/password-update', function () {
+    let organizationId;
+    let options;
+
+    beforeEach(async function () {
+      organizationId = databaseBuilder.factory.buildOrganization({ type: 'SCO', isManagingStudents: true }).id;
+      const userId = databaseBuilder.factory.buildUser().id;
+      databaseBuilder.factory.buildMembership({ organizationId, userId });
+
+      await databaseBuilder.commit();
+
+      options = {
+        method: 'POST',
+        url: '/api/schooling-registration-dependent-users/password-update',
+        headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
+        payload: {
+          data: {
+            attributes: {
+              'organization-id': organizationId,
+              'schooling-registration-id': null,
+            },
+          },
+        },
+      };
+    });
+
+    it('should return a 200 status after having successfully updated the password', async function () {
+      // given
+      const userId = databaseBuilder.factory.buildUser.withRawPassword().id;
+      const organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({
+        organizationId,
+        userId,
+      }).id;
+      options.payload.data.attributes['schooling-registration-id'] = organizationLearnerId;
+
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+
+    it('should return a 404 status when organizationLearner does not exist', async function () {
+      // given
+      options.payload.data.attributes['schooling-registration-id'] = 1;
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(404);
+    });
+
+    it("should return a 404 status when organizationLearner's userId does not exist", async function () {
+      // given
+      const organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({
+        organizationId,
+        userId: null,
+      }).id;
+      options.payload.data.attributes['schooling-registration-id'] = organizationLearnerId;
+
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(404);
+    });
+
+    it('should return a 403 status when user does not belong to the same organization as organizationLearner', async function () {
+      // given
+      const wrongOrganization = databaseBuilder.factory.buildOrganization();
+      const organizationLearnerWithWrongOrganization = databaseBuilder.factory.buildOrganizationLearner({
+        organizationId: wrongOrganization.id,
+      });
+      await databaseBuilder.commit();
+
+      options.payload.data.attributes['schooling-registration-id'] = organizationLearnerWithWrongOrganization.id;
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(403);
+    });
+  });
+
+  describe('POST /api/sco-organization-learners/password-update', function () {
+    let organizationId;
+    let options;
+
+    beforeEach(async function () {
+      organizationId = databaseBuilder.factory.buildOrganization({ type: 'SCO', isManagingStudents: true }).id;
+      const userId = databaseBuilder.factory.buildUser().id;
+      databaseBuilder.factory.buildMembership({ organizationId, userId });
+
+      await databaseBuilder.commit();
+
+      options = {
+        method: 'POST',
+        url: '/api/sco-organization-learners/password-update',
+        headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
+        payload: {
+          data: {
+            attributes: {
+              'organization-id': organizationId,
+              'organization-learner-id': null,
+            },
+          },
+        },
+      };
+    });
+
+    it('should return a 200 status after having successfully updated the password', async function () {
+      // given
+      const userId = databaseBuilder.factory.buildUser.withRawPassword().id;
+      const organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({
+        organizationId,
+        userId,
+      }).id;
+      options.payload.data.attributes['organization-learner-id'] = organizationLearnerId;
+
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+
+    it('should return a 404 status when organizationLearner does not exist', async function () {
+      // given
+      options.payload.data.attributes['organization-learner-id'] = 1;
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(404);
+    });
+
+    it("should return a 404 status when organizationLearner's userId does not exist", async function () {
+      // given
+      const organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({
+        organizationId,
+        userId: null,
+      }).id;
+      options.payload.data.attributes['organization-learner-id'] = organizationLearnerId;
+
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(404);
+    });
+
+    it('should return a 403 status when user does not belong to the same organization as organizationLearner', async function () {
+      // given
+      const wrongOrganization = databaseBuilder.factory.buildOrganization();
+      const organizationLearnerWithWrongOrganization = databaseBuilder.factory.buildOrganizationLearner({
+        organizationId: wrongOrganization.id,
+      });
+      await databaseBuilder.commit();
+
+      options.payload.data.attributes['organization-learner-id'] = organizationLearnerWithWrongOrganization.id;
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(403);
+    });
+  });
 });
