+++ api/tests/integration/application/organization-learner-dependent-users/index_test.js
@@ -14,33 +14,8 @@ describe('Integration | Application | Route | organization-learner-dependent-use
     sinon
       .stub(organizationLearnerDependentUserController, 'generateUsernameWithTemporaryPassword')
       .callsFake((request, h) => h.response('ok').code(200));
-    sinon
-      .stub(organizationLearnerDependentUserController, 'updatePassword')
-      .callsFake((request, h) => h.response('ok').code(200));
     httpTestServer = new HttpTestServer();
     await httpTestServer.register(moduleUnderTest);
-  });
-
-  describe('POST /api/schooling-registration-dependent-users/password-update', function () {
-    it('should succeed', async function () {
-      // given
-      const method = 'POST';
-      const url = '/api/schooling-registration-dependent-users/password-update';
-      const payload = {
-        data: {
-          attributes: {
-            'schooling-registration-id': 1,
-            'organization-id': 3,
-          },
-        },
-      };
-
-      // when
-      const response = await httpTestServer.request(method, url, payload);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
   });
 
   describe('POST /api/schooling-registration-dependent-users/generate-username-password', function () {
