+++ api/tests/integration/application/organization-learner-dependent-users/organization-learner-dependent-user-controller_test.js
@@ -5,8 +5,6 @@ const moduleUnderTest = require('../../../../lib/application/organization-learne
 const usecases = require('../../../../lib/domain/usecases');
 const securityPreHandlers = require('../../../../lib/application/security-pre-handlers');
 const {
-  NotFoundError,
-  UserNotAuthorizedToUpdatePasswordError,
   UserNotAuthorizedToGenerateUsernamePasswordError,
 } = require('../../../../lib/domain/errors');
 
@@ -16,7 +14,6 @@ describe('Integration | Application | organization-learner-dependent-users | org
 
   beforeEach(async function () {
     sandbox = sinon.createSandbox();
-    sandbox.stub(usecases, 'updateOrganizationLearnerDependentUserPassword').rejects(new Error('not expected error'));
     sandbox.stub(usecases, 'generateUsernameWithTemporaryPassword').resolves();
     sandbox.stub(securityPreHandlers, 'checkUserBelongsToScoOrganizationAndManagesStudents');
     httpTestServer = new HttpTestServer();
@@ -76,82 +73,6 @@ describe('Integration | Application | organization-learner-dependent-users | org
           const response = await httpTestServer.request(
             'POST',
             '/api/schooling-registration-dependent-users/generate-username-password',
-            payload,
-            auth
-          );
-
-          // then
-          expect(response.statusCode).to.equal(403);
-        });
-      });
-    });
-  });
-
-  describe('#updatePassword', function () {
-    const payload = { data: { attributes: {} } };
-    const auth = { credentials: {}, strategy: {} };
-    const generatedPassword = 'Passw0rd';
-
-    beforeEach(function () {
-      securityPreHandlers.checkUserBelongsToScoOrganizationAndManagesStudents.callsFake((request, h) =>
-        h.response(true)
-      );
-
-      payload.data.attributes = {
-        'schooling-registration-id': 1,
-        'organization-id': 3,
-      };
-
-      auth.credentials.userId = domainBuilder.buildUser().id;
-    });
-
-    context('Success cases', function () {
-      it('should return an HTTP response with status code 200', async function () {
-        // given
-        usecases.updateOrganizationLearnerDependentUserPassword.resolves(generatedPassword);
-
-        // when
-        const response = await httpTestServer.request(
-          'POST',
-          '/api/schooling-registration-dependent-users/password-update',
-          payload,
-          auth
-        );
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        expect(response.result.data.attributes['generated-password']).to.equal(generatedPassword);
-      });
-    });
-
-    context('Error cases', function () {
-      context('when a NotFoundError is thrown', function () {
-        it('should resolve a 404 HTTP response', async function () {
-          // given
-          usecases.updateOrganizationLearnerDependentUserPassword.rejects(new NotFoundError());
-
-          // when
-          const response = await httpTestServer.request(
-            'POST',
-            '/api/schooling-registration-dependent-users/password-update',
-            payload,
-            auth
-          );
-
-          // then
-          expect(response.statusCode).to.equal(404);
-        });
-      });
-
-      context('when a UserNotAuthorizedToUpdatePasswordError is thrown', function () {
-        it('should resolve a 403 HTTP response', async function () {
-          // given
-          usecases.updateOrganizationLearnerDependentUserPassword.rejects(new UserNotAuthorizedToUpdatePasswordError());
-
-          // when
-          const response = await httpTestServer.request(
-            'POST',
-            '/api/schooling-registration-dependent-users/password-update',
             payload,
             auth
           );
