+++ api/tests/unit/infrastructure/serializers/jsonapi/sco-organization-learner-serializer_test.js
@@ -86,4 +86,28 @@ describe('Unit | Serializer | JSONAPI | sco-organization-learner-serializer', fu
       expect(json).to.deep.equal(expectedSerializedOrganizationLearner);
     });
   });
+
+  describe('#serializeCredentialsForDependent', function () {
+    it('should convert into JSON API data', function () {
+      // given
+      const organizationLearner = {
+        generatedPassword: 'generated passw0rd',
+      };
+
+      const expectedSerializedOrganizationLearner = {
+        data: {
+          attributes: {
+            'generated-password': organizationLearner.generatedPassword,
+          },
+          type: 'dependent-users',
+        },
+      };
+
+      // when
+      const json = serializer.serializeCredentialsForDependent(organizationLearner);
+
+      // then
+      expect(json).to.deep.equal(expectedSerializedOrganizationLearner);
+    });
+  });
 });
